requirements.txt
Flask
PyMuPDF==1.22.0
python-docx
